Moved metadata to the crashinfo where it belongs
