+++ crashtracker/src/crash_info.rs
@@ -3,7 +3,7 @@
 
 use crate::stacktrace::StackFrame;
 use crate::telemetry::TelemetryCrashUploader;
-use crate::{CrashtrackerConfiguration, CrashtrackerMetadata};
+use crate::CrashtrackerConfiguration;
 use anyhow::Context;
 #[cfg(unix)]
 use blazesym::symbolize::{Process, Source, Symbolizer};
@@ -14,6 +14,31 @@ use std::io::BufRead;
 use std::time::Duration;
 use std::{collections::HashMap, fs::File, io::BufReader};
 use uuid::Uuid;
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct CrashtrackerMetadata {
+    pub profiling_library_name: String,
+    pub profiling_library_version: String,
+    pub family: String,
+    // Should include "service", "environment", etc
+    pub tags: Vec<Tag>,
+}
+
+impl CrashtrackerMetadata {
+    pub fn new(
+        profiling_library_name: String,
+        profiling_library_version: String,
+        family: String,
+        tags: Vec<Tag>,
+    ) -> Self {
+        Self {
+            profiling_library_name,
+            profiling_library_version,
+            family,
+            tags,
+        }
+    }
+}
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct SigInfo {
